added mock tests for quivrqarag
